Shopping cart
Shopping cart working

+++ logout.php
@@ -5,16 +5,21 @@
 	//If user is logged in...
 	if (isset($_SESSION['username'])) {
 
-		//Connect to DB
-    	$conn = connect_db();
+		if (isset($_SESSION['cid'])) {
+			//Connect to DB
+	    	$conn = connect_db();
 
-    	//Delete current shopping cart
-    	$query = "delete from cart where uid='" . $_SESSION['uid'] . "' and status='current';";
+	    	//Delete current shopping cart
+	    	$query = "delete from cart where uid='" . $_SESSION['uid'] . "' and status='current';";
 
-    	//If query fails print error and die
-		if (!mysqli_query($conn, $query)) {
-			echo "FAILURE: SQL-operation failed";
-			die();
+	    	//If query fails print error and die
+			if (!mysqli_query($conn, $query)) {
+				echo "FAILURE: SQL-operation failed";
+				die();
+			}
+
+			//Close DB connection
+			mysqli_close($conn);
 		}
 
 		//...destroy all session variables and session
